isolated LanguageToggle and ThemeToggle components
