+++ src/modules/LandingPage/landingPage.module.css
@@ -21,7 +21,7 @@
 }
 
 .logo-text {
-    font-family: Montserrat;
+    font-family: Poppins;
     font-size: 24px;
     font-weight: 600;
     color: var(--typography-headers);
